[feature] 장소 평점/리뷰 POST 기능 구현 (#55)
* feature : 장소 평점/리뷰 기능 구현

* feature : 장소 별 총 평점 계산 로직 추가

* refactor : 코스 상세조회 평점 관련 로직 수정

+++ src/main/java/umc/catchy/domain/course/dto/response/CourseInfoResponse.java
@@ -20,7 +20,7 @@ public class CourseInfoResponse {
         String courseName;
         String courseDescription;
         CourseType courseType;
-        Float rating;
+        Double rating;
         Integer reviewCount;
         String recommendTime;
         Long participantsNumber;
